change slippage tolerance to be a slider
